refactor(extensions/podman): move types to dedicated file
this is needed to refactor further without UT error
